Criando melhorias no sistema

+++ controllers/tagControllers.js
@@ -1,6 +1,7 @@
 import Tag from "../models/tagModel.js"
 
 export const getTags = async (req, res) => {
+  console.log(req.user)
   try {
     const { dataInicial, dataFinal, tipo, area, maquina } = req.query
     // console.log(area, maquina)
@@ -34,6 +35,7 @@ export const getTags = async (req, res) => {
 
 export const createTags = async (req, res) => {
   console.log(req.body)
+  console.log(req.user)
   const { tipo, area, maquina, subconjunto, conteudo } = req.body
   let data = new Date()
   let dataVence = new Date()
@@ -49,7 +51,8 @@ export const createTags = async (req, res) => {
     conteudo,
     criadaEm: data,
     venceEm: dataVence,
-    encerradaEm: data
+    encerradaEm: data,
+    userId: req.user.id
   })
   res.status(201).json({ message: 'Create Tag.' });
 }
@@ -75,6 +78,13 @@ export const deleteTag = async (req, res) => {
     if (!tag) {
       throw new Error("Etiqueta não encontrada");
     }
+
+    console.log(req.user)
+
+    if(tag.userId.toString() !== req.user.id) {
+      throw new Error("Você não pode excluir etiqueta que não é sua!");
+    }
+
     await Tag.findByIdAndDelete(id)
     res.status(200).json(tag);
   } catch (err) {
